requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
pyyaml==6.0
uritemplate==4.1.1
django-elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl-drf==0.22.5
pytest>=7.0.0,<8.0.0
pytest-django>=4.0.0,<5.0.0